# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My Disney Characters Project CRUD",
    "description": "My User Project Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:4000",
  "basePath": "/",
  "tags": [
    {
      "name": "Characters",
      "description": "API for users in the system"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],

  "paths": {
    "/characters": {
      "get": {
        "tags": ["Characters"],
        "summary": "Get all Characters in the system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Character"
            }
          }
        }
      }
    },
    "/character": {
      "post": {
        "tags": ["Characters"],
        "description": "Create a new character in the system",
        "parameters": [
          {
            "name": "character",
            "in": "body",
            "description": "Character that we want to create",
            "schema": {
              "$ref": "#/definitions/Character"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "New character is created",
            "schema": {
              "$ref": "#/definitions/Character"
            }
          }
        }
      }
    },
    "/character/{id}": {
      "delete": {
        "tags": ["Characters"],
        "summary": "Delete a character by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the character to delete",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Character deleted successfully"
          },
          "404": {
            "description": "Character not found"
          }
        }
      },
      "put": {
        "tags": ["Characters"],
        "summary": "Update a character by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the character to update",
            "required": true,
            "type": "integer"
          },
          {
            "name": "character",
            "in": "body",
            "description": "Updated character data",
            "schema": {
              "$ref": "#/definitions/Character"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Character updated successfully"
          },
          "404": {
            "description": "Character not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Character": {
      "required": ["id", "name"],
      "properties": {
        "id": {
          "type": "integer",
          "uniqueItems": true
        },
        "name": {
          "type": "string"
        }
      }
    }
  }
}